testing/web-platform/tests/workers/WorkerGlobalScope_ErrorEvent_message.htm

<!DOCTYPE html>
<html>
<head>
<title> WorkerGlobalScope onerror event handler argument: message </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>
    
    var ErrorMessage = "Error Message";
    var FileName = './support/ErrorEvent.js';
    var description = "WorkerGlobalScope onerror event handler 1st argument: message";
    
    var t = async_test("Test Description: " + description);
                  
    var worker = new Worker(FileName);

    worker.onmessage = t.step_func(function(evt)
    {
        var err = evt.data;
        assert_not_equals(err.message, undefined);
        assert_not_equals(err.message.indexOf(ErrorMessage), -1);
        t.done();        
    });
    
    worker.postMessage(ErrorMessage);
</script>
</body>
</html>
